+++ ReleaseNotes.md
@@ -5,6 +5,8 @@
 ### Features
 - cta/CTA#1001 Maximum file size is now defined by VO instead of globally.
 
+- cta/CTA#1019 New command `cta-readtp` allows reading files from tape and verifying their checksum
+
 # v4.1-1
 
 ## Summary